+++ molo/core/models.py
@@ -6,7 +6,7 @@ from django.db import models
 from django.utils.translation import ugettext_lazy as _
 from django.utils.translation import get_language_from_request
 from django.shortcuts import redirect
-from django.db.models.signals import pre_delete
+from django.db.models.signals import pre_delete, post_delete
 from django.dispatch import receiver
 
 from taggit.models import TaggedItemBase
@@ -610,9 +610,25 @@ class FooterPage(ArticlePage):
 
 FooterPage.content_panels = ArticlePage.content_panels
 
+pages_to_delete = []
+
 
 @receiver(pre_delete, sender=Page)
 def on_page_delete(sender, instance, *a, **kw):
-    for translation in PageTranslation.objects.filter(page=instance):
-        translation.translated_page.delete()
-        translation.delete()
+    ids = PageTranslation.objects.filter(
+        page=instance).values_list('translated_page__id')
+    pages_to_delete.extend(Page.objects.filter(id__in=ids))
+
+
+@receiver(post_delete, sender=Page)
+def on_post_page_delete(sender, instance, *a, **kw):
+    # When we try to delete a translated page in our pre_delete, wagtail
+    # pre_delete function would want to get the same page too, but since we
+    # have already deleted it, wagtail would not be able to find it, therefore
+    # we have to get the translated page in our pre_delete and use a global
+    # variable to store it and pass it into the post_delete and remove it here
+    for p in pages_to_delete:
+        p.delete()
+
+    global pages_to_delete
+    del pages_to_delete[:]